backend/requirements.txt
alembic==1.14.0
bcrypt==4.2.0
blinker==1.8.2
blueprint==3.4.2
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
Deprecated==1.2.15
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-exporter-prometheus==0.50b0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-flask==0.50b0
opentelemetry-instrumentation-logging==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-wsgi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
packaging==24.1
prometheus_client==0.21.1
protobuf==5.29.2
psycopg2-binary==2.9.10
pytube==15.0.0
pytubefix==8.7.0
requests==2.32.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.17.0
zipp==3.21.0
